Adding the maim HTML structure
In visual studio code I added the main HTML structure to start with the project. In visual studio code there is a shortcut to add the main HTML structure.
Just type ! and press ENTER, visual studio code add automaticaly the main HTML code. That is the DOCTYPE, the html root element, the encoding, the head and the body.